refactor(mentor): Move time-related functions to utils/time-functions directory
