+++ src/components/LinkCard/constants.js
@@ -7,20 +7,6 @@ const STYLE_LINK_CARD_CREATED_CLASS_NAME = 'link-card-created';
 const STYLE_THUMBNAIL_CONTAINER_CLASS_NAME = 'thumbnail-container';
 const STYLE_THUMBNAIL_CLASS_NAME = 'thumbnail';
 
-const MILLISECONDS_IN_YEAR = 31536000000;
-const MILLISECONDS_IN_MONTH = 2592000000;
-const MILLISECONDS_IN_DAY = 86400000;
-const MILLISECONDS_IN_HOUR = 3600000;
-const MILLISECONDS_IN_MINUTE = 60000;
-
-const INTERVALS = [
-  { label: 'year', divisor: MILLISECONDS_IN_YEAR },
-  { label: 'month', divisor: MILLISECONDS_IN_MONTH },
-  { label: 'day', divisor: MILLISECONDS_IN_DAY },
-  { label: 'hour', divisor: MILLISECONDS_IN_HOUR },
-  { label: 'minute', divisor: MILLISECONDS_IN_MINUTE },
-];
-
 export {
   STYLE_LINK_CARD_CLASS_NAME,
   STYLE_LINK_CARD_INFO_CLASS_NAME,
@@ -30,10 +16,4 @@ export {
   STYLE_LINK_CARD_CREATED_CLASS_NAME,
   STYLE_THUMBNAIL_CONTAINER_CLASS_NAME,
   STYLE_THUMBNAIL_CLASS_NAME,
-  MILLISECONDS_IN_YEAR,
-  MILLISECONDS_IN_MONTH,
-  MILLISECONDS_IN_DAY,
-  MILLISECONDS_IN_HOUR,
-  MILLISECONDS_IN_MINUTE,
-  INTERVALS,
 };
