+++ src/utils/time-functions/formatDate.js
@@ -1,0 +1,3 @@
+export default function formatDate(date) {
+  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
+}
